requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==2.7.0
joblib==1.4.0